% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{as_numeric}
\usage{
as_numeric(x, na.strings = NULL, type = class(x))
}
\arguments{
\item{x}{vector}

\item{na.strings}{Fehlende Werte}

\item{type}{factor, numeric, time}
}
\description{
String und Factor zu Numeric.
}
\examples{

x <- c("0","-3","+5"," 5.6","   3,5  ","3,,4","3-5","uk","na",-77,-99,-999)
f<- factor(c(4,  16, 8,  1,  4,  1,  2,  4,  2,  2,  -77,  4 ))

data.frame(x=x,
           x_as_num=as_numeric(x, na.strings = c("-77", "-99")),
           
           f=f,
           f_as_num=as_numeric(f, na.strings = c("-77", "-99"))) 
}
