% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{clean_names
Quelle https://drdoane.com/clean-consistent-column-names/}
\usage{
clean_names(.data, tolower = FALSE, unique = TRUE,
  abbreviate = FALSE, minlength = 4)
}
\arguments{
\item{.data}{data.frame oder character}

\item{tolower}{alles zu kleinbuschstaben}

\item{unique}{eindeitige namen}

\item{abbreviate, minlength}{Abbkürzung}
}
\value{
die selbe Klasse wie dr Input
}
\description{
clean_names
Quelle https://drdoane.com/clean-consistent-column-names/
}
\examples{


clean_names(tibble::tibble("Ã–li"=1:3, "p-k"=1:3, "95\%-CI"=4:6) )

clean_names(
  c(
    "  a", "a  ", "a \%", "a",
    "$a", "$$$a", "GDP ($)",
    "GDP (us$)", "a (#)", "a & b",
    "#", "$", "a_cnt", "Aa&Bb",
    "camelCasePhrases", "AlphaBetaGamma",
    "Alpha       Beta", "Beta  !!! Gamma",
    "a + b", "a - b", "a * b", "Ã–sel"
  ), abbreviate=TRUE
)
}
