% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{CleanUp}
\alias{CleanUp}
\alias{CleanUp.tbl_df}
\alias{cleanup_names_encoding}
\alias{CleanUp.default}
\alias{CleanUp_factor}
\title{CleanUp}
\usage{
CleanUp(data, ...)

\method{CleanUp}{tbl_df}(data, ...)

cleanup_names_encoding(data, from = "UTF8", to = "latin1")

\method{CleanUp}{default}(
  data = NA,
  uselabels = FALSE,
  na.strings = NULL,
  force.numeric = FALSE,
  encoding = TRUE,
  sep = ".",
  variable.labels = attributes(data)$variable.labels,
  variable.names = names(data),
  ...
)

CleanUp_factor(data, ...)
}
\arguments{
\item{data}{Date.Frame}

\item{...}{extra arguments}

\item{from}{encoding from UTF8}

\item{to}{encoding to latine1}

\item{uselabels}{Labels labels   FALSE}

\item{na.strings}{Fehlende Werte}

\item{force.numeric}{alles zu Zahlen alles zu Zahlen konvertieren}

\item{encoding}{Stichwort sonderzeichen  "UTF-8"}

\item{sep}{Trennzeichen variable.labels   noch nicht getestet  attributes(data)$variable.labels,}

\item{variable.labels}{Variable Labels  Import von LimeSurvy}

\item{variable.names}{Variable Namen}
}
\description{
Bereinigung der Labels
\code{CleanUp_factor()} Die Labels von tibble::tibble werden ja als labelled -Attribut abgelegt und durch cleanup werden sie f?r APA2 lesber clenaUp.tibble::tibble

Die \code{CleanUp.tibble::tibble()} bereinigt die Faktoren da die Auswertung
direkt oft nicht funktioniert

CleanUp_factor bereinigen von haven Label
}
