% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{cleanup_NA}
\alias{cleanup_NA}
\title{cleanup_NA}
\usage{
cleanup_NA(obj, na.strings = NULL, force.numeric = FALSE)
}
\arguments{
\item{obj}{data.frame}

\item{na.strings}{na als character}

\item{force.numeric}{alles zu Nummern}
}
\value{
data.frame
}
\description{
cleanup_NA
}
