% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codebook.R
\name{Codebook}
\alias{Codebook}
\title{Codebook}
\usage{
Codebook(data, abbreviate = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{abbreviate}{Abkuerzen del Levels}
}
\description{
Extrahier die Daten-Struktur
}
