% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cs2.R
\name{Cs2}
\alias{Cs2}
\title{Character strings from unquoted names}
\usage{
Cs2(...)
}
\arguments{
\item{...}{Vektor oder String oder Formula}
}
\value{
Vektor
}
\description{
Erweiterung der  Hmisc Cs-Funktion um trim.ws. Erlaubt Variablen mit Komma oder Plus(+) oder Leerzeichen abzuternnen.
}
\examples{
Cs2(sd,fr,fg)
Cs2(sd,fr,fg, "hju nh")
Cs2("  Hallo Welt ")
Cs2("~ sd +    fr + fg")
}
\seealso{
\link{Cs}
}
