% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapply2.R
\name{Dapply}
\alias{Dapply}
\alias{dapply2}
\alias{Dapply.formula}
\alias{Dapply.data.frame}
\title{Dapply}
\usage{
Dapply(x, ...)

dapply2(
  data,
  fun = function(x) as.numeric(x),
  stringsAsFactors = default.stringsAsFactors(),
  ...
)

\method{Dapply}{formula}(
  x,
  data,
  fun = function(xx) {     as.numeric(xx) },
  stringsAsFactors = default.stringsAsFactors(),
  ...
)

\method{Dapply}{data.frame}(
  data,
  ...,
  fun = function(x) as.numeric(x),
  stringsAsFactors = default.stringsAsFactors()
)
}
\arguments{
\item{x}{Objekt}

\item{...}{Weitere Argumente an llply}

\item{data}{Data.frame}

\item{fun}{funktion function(x) as.numeric(x)}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}
}
\value{
data.frame
}
\description{
Dapply, dapply2 ist  plyr::llply()
}
\examples{
 
df1 <- Label(data.frame(
  month = rep(1:3, 2),
  student = rep(c("Amy", "Bob"), each = 3),
  A = c(9, 7, 6, 8, 6, 9),
  B = c(6, 7, 8, 5, 6, 7)
),
A = "Deutsch",
B = "Mathe")

rs1 <- Dapply(~ A + B,
              df1,
              cut,
              breaks = 3,
              labels = c(1:3))
              
}
