% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter2.R
\name{Filter2}
\alias{Filter2}
\alias{subset2}
\alias{Select_case}
\alias{Drop_NA}
\alias{Drop_case}
\title{Subsetting Vectors, Matrices and Data Frames}
\usage{
subset2(data, ...)

Select_case(data, ..., output = TRUE)

Filter2(data, ...)

Drop_NA(data, ..., output = TRUE)

Drop_case(data, subset = NULL, output = TRUE)
}
\arguments{
\item{...}{further arguments to be passed to or from other methods.}

\item{subset}{Ausgewählte Fälle Position als Zahl}

\item{x}{object to be subsetted.}
}
\value{
data.frame
}
\description{
Select_case, subset2 \code{base::subset()}

Filter2 \code{dplyr::filter()}

Drop_NA    \code{tidyr::drop_na()}

Drop_case   \code{edata[-subset, ] }
}
\examples{

 dat <- data.frame(
nr= 1:5,
sex = factor(c(1, 2, 1, 2, 1), 1:3, c("m", "f", "t")),
treatment = c("A", "A", "B", "B", "A"),
m1 = c(1, NA, 2, 1, 1),
m2 =  c(1, 5, NA, 1, 1),
m3 = c(2, 1, 5, 1, 1),
m4 = c(3, 2,2, 2, 1),
m5 = c(5, 2, 3, 1, 2),
m6 = c(4, 1, 4, 1, 1)
)

dat <- Label(
  dat,
  sex = "Geschlecht",
  treatment = "Behandlung",
  m1 = "Cohlesterin",
  m2 = "Billirubin"
)

dat <- tibble::as_tibble(dat)


subset2(dat, treatment == "A" & sex=="m" )
Select_case(dat, treatment == "A" & sex=="m", output=FALSE)
Filter2(dat, treatment == "A" & sex=="m")

Drop_NA(dat, m1)
# tidyr::drop_na(dat,m1)
Drop_NA(dat)
Drop_case(dat, c(2:4))


}
