% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{Filter_Subset}
\alias{Filter_Subset}
\alias{Drop_NA}
\alias{Drop_case}
\alias{Select_case}
\title{Filter und Subset}
\usage{
Drop_NA(data, ...)

Drop_case(data, subset = NULL, ...)

Select_case(data, ...)
}
\arguments{
\item{subset}{Ausgewählte Fälle}
}
\description{
\code{Drop_NA} entfernt Fälle mit NAs mit \code{dplyr::filter}

\code{Drop_case} entfernt Fälle arbeitet auch wie subset()

\code{Drop_case} entfernt Fälle arbeitet auch wie subset()
}
\examples{
data<- data.frame(g= gl(2, 8, labels = c("Control", "Treat")),
a=rpois(16, 5),
b=rep(1:2,8),
c=rnorm(16))

data$d<-rep(c("", "b", "c"), 6)[1:16]
data[2,2]<- NA
str(data<-upData2(data, labels = c(g="Gruppe" , a="A", b="B") ))
str(Drop_NA(data))

str(Drop_case(data, 1))

#-  Aber das geht auch
APA2(~., Drop_case(data, 1)  )

APA2(~., dplyr::slice(data, -1))


#' str(Select_case(data, g=="Control"))

}
