% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.R
\name{delet_labels}
\alias{delet_labels}
\alias{set_label}
\alias{get_label}
\alias{GetLabelOrName}
\title{get_label}
\usage{
delet_labels(data)

set_label(data, labels = NULL)

get_label(x)

GetLabelOrName(x, pattern = NULL, replacement = " ")
}
\arguments{
\item{labels}{Labels}

\item{x}{data.drame}

\item{pattern}{pattern = "[\\._]+"}

\item{replacement}{Leerzeichen}
}
\value{
Namend Character String
}
\description{
DeletLabels:  Loeschen des Attributs label

Labels ohne die unit bei GetLabelOrName weid auch die unit verwendet.

Extrahiert die Labels
}
\examples{

# GetLabelOrName(data, pattern="störender string")

}
