% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Daten Importieren}
\usage{
GetData(File = NA, na.strings = NULL, force.numeric = FALSE,
  Tabel_Expand = FALSE, id.vars = 1, value = "value",
  Data_info = date(), sep = ";", quote = "\\"", dec = ".",
  user_na = FALSE, reencode = FALSE, ...)
}
\arguments{
\item{File}{Demo.xla, Demo.sav, Demo.csv  - ist ein Pfad zu einem csv, xls, sav oder Rdata Datensatz oder ein String in Textformat dierekt im R-File.}

\item{na.strings}{c(NA,9999,"") - Fehlende Werte}

\item{force.numeric}{logical FALSE}

\item{Tabel_Expand}{logical FALSE - Tabellen mit haufigkeiten werden als Dataframe im long-Format ausgegeben}

\item{id.vars}{nur mit Tabel_Expand  - Nummer und Name der ID-Variablen bei Tabel_Expand default ist 1.}

\item{value}{nur mit Tabel_Expand  - Name der output-variable bei Tabel_Expand.}

\item{Data_info}{Data_info = date(),}

\item{sep}{Lesen der csv- Files = ";",}

\item{quote}{Lesen der csv- Files = "\"",}

\item{dec}{Lesen der csv- Files = ".",}

\item{user_na}{If TRUE variables with user defined missing will be read into labelled_spss objects. If FALSE, the default, user-defined missings will be converted to NA.}

\item{reencode}{UTF-8 = FALSE,}

\item{...}{Argumente fuer spss und csv. Bei SPSS-Files  kann die Zeichencodierung mit  \code{reencode ="UTF-8"} geaendert werden.}
}
\description{
Ladet verschiedene Dateiformate von csv bis sav.  Tabellen im Text-Format koennen direkt gelesen werden.
Zurueckgegeben wird ein Hmisc data.frame mit Levels und Labels siehe \link{upData}.
\subsection{CSV-Files}{Liest CSV-Files siehe \link{read.csv}   uebergeben werden Trennzeichen, Komma, usw.
  }
\subsection{Text direkt im R-File}{
 Es lassen sich direkt Daten als Text einlesen dabei kann mit \code{Tabel_Expand = TRUE, id.vars = 1:2} gesteuert
 werden ob eine Kreuztabelle aufgedroeselt werden soll.
  }
\subsection{SPSS}{
Mit \code{GetData("Dummy.sav", reencode="UTF-8")} lassen sich SPSS-Files einlesen. Bei SPSS
Dateien (sav und por) werden leere Labels
automatisch bereinigt dH sie werden durch die Number des Faktors ersaetzt. Probleme
machen lange Text-Variablen im SPSS-File, die sollten im Falle geloescht werden. Weitere
Info unter  \link{spss.get} und \link{read.spss}.
Neu ist das  die Default-Option   spss.get() is  um mit haven zu Arbeitenist als Parameter  \code{type=2}
zu ?bergeben. Bei hven werden zus?tzlich die Faktoren bereinigt.
  }
\subsection{LimeSurvy}{
 Hier muss eine Liste uebergeben werden die die Filenamen beinhaltet also \code{GetData(list("file.R","file.txt"))}.
 Das erste File ist das R-File mit den Labels das zweite die Daten. Weitere moegliche Parameter sind die Zeichencodierung.
 }
\subsection{RData}{
 Hier erfolgt das Einlesen einfach ueber \code{load(File, ...)}, besser ist es daher direkt den Befehl zu verwenden.
 }
}
