% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLabelOrName.R
\name{GetLabelOrName}
\alias{GetLabelOrName}
\title{GetLabelOrName Hmisc labels}
\usage{
GetLabelOrName(x, pattern = NULL, replacement = " ")
}
\arguments{
\item{x}{Data.frame}

\item{pattern}{pattern = "[\\._]+"}

\item{replacement}{Leerzeichen}
}
\description{
Extrahiert die Labels
}
\examples{
# GetLabelOrName(data, pattern="störender string")

}
\author{
Wolfgang Peter
}
