% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.R
\name{Label}
\alias{Label}
\alias{delet_labels}
\alias{set_label}
\alias{get_label}
\alias{wrap_label}
\alias{GetLabelOrName}
\alias{label_data_frame}
\title{Label}
\usage{
Label(data, ...)

delet_labels(data)

set_label(data, labels = NULL)

get_label(data, include.units = FALSE)

wrap_label(x, width = 20, sep = NULL, pattern = "_", replacement = " ")

GetLabelOrName(data, include.units = TRUE)

label_data_frame(data, labels)
}
\arguments{
\item{data}{data.frame}

\item{...}{Labels in der Form a="Hallo, b="Welt"}

\item{labels}{labels als named vector}

\item{include.units}{Einheiten}

\item{x}{data.frame or string}

\item{width}{laenge 20}

\item{sep}{default newline}

\item{pattern, replacement}{an gsub}
}
\description{
Setzen des attr(data, "label")

delet_labels:  Loeschen aller Attributs label

set_label:  Setzen  der Attributs label

get_label und GetLabelOrName:  Abrufen  der Attributs label

wrap_label kuerzt die Labels fur Grafiken.

Intern wenn mit get_label nur die Kopie wiederhergestellt wird.
}
\examples{

df <- data.frame(
BMI=c(1,2,3,1,2,3),
WHtR= gl(2,3, labels =c("Amy", "Bob")),
WHtR_1=c(9,7,6,8,6,9),
bildprof=c(6,7,8,5,6,7)
)

DF<-
  Label(df, BMI = "Body-Mass-Index",
        WHtR =  "Waist-Height-Ratio",
        WHtR_1 ="Waist-Height-Ratio"
  )

DF$BMI<- units::set_units(DF$BMI, kg/m2)

get_label(DF)
get_label(DF, include.units=TRUE)
DF<- set_label(DF, c(bildprof = "Bildungsprofil"))
get_label(DF)
DF<- delet_labels(DF)
get_label(DF)

DF<- wrap_label(DF)
get_label(DF)

}
