% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melt2.R
\name{Melt2}
\alias{Melt2}
\alias{Melt2.formula}
\alias{Melt2.data.frame}
\alias{Melt2.default}
\alias{melt2}
\title{Umformen (reshape)}
\usage{
Melt2(x, ...)

\method{Melt2}{formula}(x, data, key = "variable", value = "value",
  subset, na.action = na.pass, X = stp25APA2::Formula_Data(x, data,
  subset, na.action), id.vars = X$xname, ...)

\method{Melt2}{data.frame}(x, ..., key = "variable", value = "value")

\method{Melt2}{default}(data, ...)

melt2(x, ..., by = "1")
}
\arguments{
\item{x}{Objekt kann Formel oder data.frame sein}

\item{...}{weitere Argumente}

\item{data}{Daten}

\item{key}{Bezeichnung der Bezeichner-Variable default ist "variable"}

\item{value}{Bezeichnung der Werte-Variable default ist "value"}

\item{subset}{nicht verwendbar}

\item{na.action}{auch nicht zu veraendern}

\item{X}{Formula-Objekt nicht ändern}

\item{id.vars}{nur bei Methode data.frame zu verwenden sonst ist hier nichts zu veraendern}

\item{by}{Gruppierung}

\item{...}{weitere Argumente an melt}
}
\value{
Dataframe in Langform
}
\description{
Umformen von einem Breit-Format nach einem Lang-Format. Melt2 und melt2 sind
Erweiterungen der reshape2::melt Funktion. Intern wird \code{melt} und \code{dcast} verwendet.

Melt2.formula: Melt2(chol0+chol1+chol6+chol12~g, hyper)

Melt2.data.frame ist wie melt zu verwenden aber die Labels werden dabei benuzt.

Melt2.default gibt alles ausser die data.frames und Formeln an reshape2::melt weiter.

melt2 ist die lazy_dots-Methode
}
\examples{
x<-Melt2(chol0+chol1+chol6+chol12~g , hyper)
aggregate( value~variable, x, mean)

#-- Melt2.data.frame--

x  <- hyper[, c("g","chol0","chol1","chol6","chol12")]
x  <- Melt2(x, id.vars=1)
# aggregate(value~variable+g, x, mean)

# Alternative aber ohne die Labels
x <- hyper  \%>\%
    tidyr::gather("time", "chol", chol0:chol12) \%>\%
    dplyr::select(g, time, chol)


 head(x<-Melt2(chol0+chol1+chol6+chol12~g , hyper))
 # APA2(~.,x)

 head( x<- hyper \%>\%  melt2(chol0,chol1,chol6,chol12, by=~g))
 #APA2(~.,x)
}
\seealso{
\link[reshape2]{melt} und \link[reshape2]{dcast}
}
