% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{read.text2}
\alias{read.text2}
\title{Read Text Lines}
\usage{
read.text2(
  string,
  na.strings = c("NA", "na"),
  sep = "\\t",
  dec = ".",
  stringsAsFactors = TRUE
)
}
\arguments{
\item{string}{character string}

\item{na.strings}{a character vector of strings which are to be interpreted as NA values.}

\item{sep}{the field separator character.}

\item{dec}{the character used in the file for decimal points.}

\item{stringsAsFactors}{}
}
\value{
data.frame
}
\description{
Read Text Lines
}
