% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recast2.R
\name{Recast2}
\alias{Recast2}
\alias{Summarise}
\title{Summarise Data}
\usage{
Recast2(Formula, data, fun = NULL, key = "variable", value = "value",
  subset, na.action = na.pass, X = stp25formula::Formula_Data(Formula,
  data, subset, na.action), id.vars = X$xname, measure.var,
  formula = NULL, labels = TRUE, margins = FALSE,
  margins_name = "gesamt", ...)

Summarise(x, .data, fun = NULL, key = "variable", value = "value",
  subset, na.action = na.pass, X = stp25formula::Formula_Data(x, .data,
  subset, na.action), id.vars = X$xname, measure.var, formula = NULL,
  labels = TRUE, margins = FALSE, margins_name = "gesamt",
  key.levels = NULL, ...)
}
\arguments{
\item{Formula}{Formula-Objekt}

\item{data}{Daten}

\item{fun}{Agregat Funktion}

\item{key, value}{Names of new key and value columns, as strings}

\item{subset, na.action}{fuer Formula na.action = na.omit, na.pass, na.exclude, na.fail}

\item{X}{Formula-Objekt}

\item{measure.var}{Name der Gemessenen Variable}

\item{formula}{zum vertauchen von Spalten Argument an dcast}

\item{labels}{sollen Hmisc levels verwendet werden}

\item{margins}{Gesamtwert bei TRUE eigene Funktion sonst dcast}

\item{...}{recast Arguments \link[reshape2]{recast}}
}
\value{
reshape objek
}
\description{
Erweiterung der recast-Funktion. Erwartet einen data.frame
mit \code{label} und verwendet die \code{melt} und \code{dcast}  Funftionen
vom Packet library(reshape2).
}
\examples{

#library(tidyverse)
#library(stp25aggregate)
#library(stp25data)

hyper1<-hyper[, c("g","chol0","chol1","chol6","chol12")]
hyper_long<- Melt2(hyper1, id.vars=1)
aggregate( value~variable, hyper_long, mean)

hyper_long<-Melt2(chol0+chol1+chol6+chol12~g,
                  hyper, "Zeit", "Cholesterin")
#-- Spread + aggragate
aggregate(Cholesterin~Zeit+g, hyper_long, mean) \%>\%
  spread(g, Cholesterin)

#- das gleiche wie aggragate
hyper_long \%>\% group_by(Zeit, g) \%>\%
  summarise(Cholesterin=mean(Cholesterin)) \%>\%
  spread(g, Cholesterin)

#-- Gather das gleiche wie oben aber ohne die Labels
hyper  \%>\%
  tidyr::gather("time", "chol", chol0:chol12) \%>\%
  dplyr::select(g, time, chol)

#-- Recast2 das gleiche wie oben
Recast2(chol0+chol1+chol6+chol12~g, hyper, mean,
        "Zeit", "Cholesterin") \%>\%
  spread(g, Cholesterin)

#-- Recast2 das gleiche wie oben nur ohne  spread
Recast2(chol0+chol1+chol6+chol12~g, hyper, mean,
        formula=variable~g)


Recast2(chol0+chol1+chol6+chol12~g, hyper,
        mean,formula=variable~g, margins=TRUE)

}
\seealso{
\link[reshape2]{recast}
}
\author{
Wolfgang Peter
}
