% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset2}
\alias{subset2}
\title{Subsetting Vectors, Matrices and Data Frames}
\usage{
subset2(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{logical expression indicating elements or rows to keep: missing values are taken as false.}

\item{select}{expression, indicating columns to select from a data frame.}

\item{drop}{passed on to [ indexing operator.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
data.frame
}
\description{
Subsetting Vectors, Matrices and Data Frames
}
\examples{


 dat <- data.frame(
sex = factor(c(1,2,1,2,1),1:3,c("m","f","t")),  treatment = c("A", "A", "B", "B", "A"),
m1 = c(1, NA, 2, 1,1),  m2 = 1:5,  m3 = 1:5,  m4 = 1:5,  m5 = 1:5,  m6 = 1:5
)
dat <- Label(dat, sex="Geschlecht",
                             treatment="Behandlung",
                             m1="Cohlesterin", m2="Billirubin")

dat<-tibble::as_tibble(dat)
subset(dat, treatment=="A")


}
