% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise.R
\name{Summarise}
\alias{Summarise}
\alias{Recast2}
\title{Summarise Data}
\usage{
Summarise(
  ...,
  fun = function(x) length(na.omit(x)),
  key = "variable",
  value = "value",
  na.action = na.pass,
  formula = NULL,
  margins = FALSE,
  margins_name = "Total"
)

Recast2(
  Formula,
  data,
  fun = function(x) length(na.omit(x)),
  key = "variable",
  value = "value",
  na.action = na.pass,
  X,
  id.vars,
  measure.var,
  formula = NULL,
  labels,
  margins = FALSE,
  margins_name = "gesamt"
)
}
\arguments{
\item{...}{An Long() daten Formeln,Variablen-Namen}

\item{fun}{function(x) length(na.omit(x)) an aggregate}

\item{key, value}{an Long bzw}

\item{na.action}{an aggregate}

\item{formula}{Zeilen/Spalten Lang/Weit}

\item{margins, margins_name}{Gesamt}
}
\value{
reshape objek
}
\description{
Summarise: ist eine  Erweiterung  aggregate()
}
\examples{

 mean3<- function(x) mean(x, round(na.rm=TRUE),2)
df <- data.frame(
  month = rep(1:3, 2),
  student = rep(c("Amy", "Bob"), each = 3),
  A = c(9, 7, 6, 8, 6, 9),
  B = c(6, 7, 8, 5, 6, 7)
)

Summarise(A + B ~ student, df )

df \%>\%  Summarise2(
  A,
  B,
  by =  ~ student+ month,
  fun = mean3,
  formula=student ~ month,
  margins = TRUE 
)
#Recast(chol0+chol1+chol6+chol12~g, hyper, mean, "Zeit", "Cholesterin")  
#Summarise(chol0+chol1+chol6+chol12~g, hyper, fun= mean,  key="Zeit", value="Cholesterin")


}
