% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{TabelToExpandDataFrame}
\alias{TabelToExpandDataFrame}
\title{Tabel To Expand Data Frame}
\usage{
TabelToExpandDataFrame(
  data,
  id.vars,
  value = "value",
  na.strings = "NA",
  as.is = FALSE,
  dec = "."
)
}
\arguments{
\item{data}{}

\item{id.vars}{}

\item{value}{}

\item{na.strings}{}

\item{as.is}{}

\item{dec}{}
}
\value{
data.frame
}
\description{
Tabel To Expand Data Frame
}
\examples{

dat <- TabelToExpandDataFrame("
sex treatment  neg  pos
f   KG          3   3
f   UG          4   5
m   KG          5   4
m   UG          4   2
",
id.vars = 1:2,
value = "befund")

xdat <- xtabs( ~ befund + sex + treatment, dat)
Wide(as.data.frame(xdat),
     befund ,
     Freq)
}
