% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XLS.R
\name{XLS}
\alias{XLS}
\title{Nummernindex aus Excel - Spaltenbezeichnung}
\usage{
XLS(
  ...,
  myLetters = c(LETTERS, unlist(lapply(LETTERS, function(abc) paste0(abc, LETTERS))))
)
}
\arguments{
\item{...}{liste mit den Spaltennamen A:BB}

\item{myLetters}{nicht zum ändern}
}
\description{
Extrahiert aus Buchstaben die Spaltennummer
}
\examples{

#strsplit("A:V", "\\:")
XLS(a, B)
XLS(a, B, c:f, g:h,i, r:z)
XLS(A:Z)
}
\seealso{
\link{strsplit}
}
\author{
Wolfgang Peter
}
