% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{CleanUp}
\alias{CleanUp}
\alias{CleanUp.tbl_df}
\alias{cleanup_names_encoding}
\alias{CleanUp.default}
\alias{CleanUp_factor}
\alias{Label}
\alias{label_data_frame}
\title{CleanUp}
\usage{
CleanUp(data, ...)

\method{CleanUp}{tbl_df}(data, ...)

cleanup_names_encoding(data, from = "UTF8", to = "latin1")

\method{CleanUp}{default}(data = NA, uselabels = FALSE, na.strings = NULL,
  force.numeric = FALSE, encoding = NULL, sep = ".",
  variable.labels = attributes(data)$variable.labels,
  variable.names = names(data), ...)

CleanUp_factor(data, ...)

Label(data, ...)

label_data_frame(data, labels = NULL)
}
\arguments{
\item{data}{Date.Frame}

\item{...}{extra arguments}

\item{from}{encoding from UTF8}

\item{to}{encoding to latine1}

\item{uselabels}{Labels}

\item{na.strings}{Fehlende Werte}

\item{force.numeric}{alles zu Zahlen}

\item{encoding}{Stichwort sonderzeichen}

\item{sep}{Trennzeichen}

\item{variable.labels}{Variable Labels}

\item{variable.names}{Variable Namen}

\item{labels}{Labels}
}
\description{
Bereinigung der Labels
\code{CleanUp_factor()} Die Labels von tbl_df werden ja als labelled -Attribut abgelegt und durch cleanup werden sie f?r APA2 lesber clenaUp.tbl_df

Die \code{CleanUp.tbl_df()} bereinigt die Faktoren da die Auswertung
direkt oft nicht funktioniert

Die Funktion \code{Label()} erstellt die
Labels die bei den Tabellen
verwendet werden. Im Unterschied zu \code{UpData2()} arbeitet diese Funktion mit
  (...) argument.
}
\examples{

#-- Label
# Steko<- GetData("Steko.sav") \%>\%
# Drop_NA(key) \%>\%
#   mutate(jahr = factor(jahr)) \%>\%
#   Label( BMI = "Body-Mass-Index",
#         WHtR =  "Waist-Height-Ratio",
#         WHtR_1 ="Waist-Height-Ratio",
#         bildprof = "Bildungsprofil",
#         jahr = "Jahr")
#
# MobilePayment \%>\%
#  select(Q2_1:Q2_4) \%>\%
#   Label("Computer", "Laptop", "Tablet", "Mobiltelefon" )

}
