% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R, R/Label.R, R/subset.R
\name{CleanUp}
\alias{CleanUp}
\alias{CleanUp.tbl_df}
\alias{cleanup_names_encoding}
\alias{CleanUp.default}
\alias{CleanUp_factor}
\alias{Label}
\alias{SetLabels}
\alias{DeletLabels}
\alias{label_data_frame}
\alias{Drop_NA}
\alias{Drop_case}
\alias{Select_case}
\title{CleanUp}
\usage{
CleanUp(data, ...)

\method{CleanUp}{tbl_df}(data, ...)

cleanup_names_encoding(data, from = "UTF8", to = "latin1")

\method{CleanUp}{default}(data = NA, uselabels = FALSE,
  na.strings = NULL, force.numeric = FALSE, encoding = TRUE,
  sep = ".", variable.labels = attributes(data)$variable.labels,
  variable.names = names(data), ...)

CleanUp_factor(data, ...)

Label(data, ...)

SetLabels(x, ...)

DeletLabels(data)

label_data_frame(data, labels = NULL)

Drop_NA(data, ..., output = TRUE)

Drop_case(data, subset = NULL, output = TRUE)

Select_case(data, ..., output = TRUE)
}
\arguments{
\item{data}{Date.Frame}

\item{...}{extra arguments}

\item{from}{encoding from UTF8}

\item{to}{encoding to latine1}

\item{uselabels}{Labels}

\item{na.strings}{Fehlende Werte}

\item{force.numeric}{alles zu Zahlen}

\item{encoding}{Stichwort sonderzeichen}

\item{sep}{Trennzeichen}

\item{variable.labels}{Variable Labels}

\item{variable.names}{Variable Namen}

\item{labels}{Labels}

\item{subset}{Ausgew채hlte F채lle Position als Zahl}
}
\description{
Bereinigung der Labels
\code{CleanUp_factor()} Die Labels von tbl_df werden ja als labelled -Attribut abgelegt und durch cleanup werden sie f?r APA2 lesber clenaUp.tbl_df

Die \code{CleanUp.tbl_df()} bereinigt die Faktoren da die Auswertung
direkt oft nicht funktioniert

CleanUp_factor bereinigen von haven Label

Die Funktion \code{Label()} erstellt die
Labels die bei den Tabellen
verwendet werden. Im Unterschied zu \code{UpData2()} arbeitet diese Funktion mit
  (...) argument.

DeletLabels:  Loeschen des Attributs label

\code{Drop_NA} entfernt F채lle mit NAs mit \code{dplyr::filter(data, ...)} wenn \code{...} leer ist ann alle also quasi na.omit

\code{Drop_case} entfernt F채lle arbeitet mit data[-subset, ] + label_data_frame()

\code{Select_case} entfernt Faelle arbeitet mit subset(data, ...) + label_data_frame()
}
\examples{

#-- Label
# Steko<- GetData("Steko.sav") \%>\%
# Drop_NA(key) \%>\%
#   mutate(jahr = factor(jahr)) \%>\%
#   Label( BMI = "Body-Mass-Index",
#         WHtR =  "Waist-Height-Ratio",
#         WHtR_1 ="Waist-Height-Ratio",
#         bildprof = "Bildungsprofil",
#         jahr = "Jahr")
#
# MobilePayment \%>\%
#  select(Q2_1:Q2_4) \%>\%
#   Label("Computer", "Laptop", "Tablet", "Mobiltelefon" )

data<- data.frame(g= gl(2, 8, labels = c("Control", "Treat")),
a=rpois(16, 5),
b=rep(1:2,8),
c=rnorm(16))

data$d<-rep(c("", "b", "c"), 6)[1:16]
data[2,2]<- NA
str(data<-upData2(data, labels = c(g="Gruppe" , a="A", b="B") ))
str(Drop_NA(data))

 tidyr::drop_na(data)
 

str(Drop_case(data, 1))

#-  Aber das geht auch
APA2(~., Drop_case(data, 1)  )

APA2(~., dplyr::slice(data, -1))


#' str(Select_case(data, g=="Control"))

}
