% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapply2.R
\name{Dapply}
\alias{Dapply}
\title{Make New Variables}
\usage{
Dapply(x, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{an dapply2}
}
\value{
dataframe
}
\description{
Apply mit formula
}
\examples{
library(tidyverse)
library(stp25aggregate)
library(stp25data)
res <- Dapply(~ a  +  gr  + gew+chol0+chol1+chol6, hyper, fun=function(x) cut(x,3))
head(res)


}
