% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapply2.R
\name{dapply2}
\alias{dapply2}
\title{Apply mit Hmisc labels}
\usage{
dapply2(.data, fun = function(x) as.numeric(x), ...)
}
\arguments{
\item{.data}{Data.frame}

\item{fun}{funktion}

\item{...}{Weitere Argumente an llply}
}
\value{
upData dataframe
}
\description{
Mischung usd der \code{llply} und \code{data.frame} -Funktion.
Erwartet einen data.frame und die Daten werdentransformiert.
Wenn keine Funktion uebergeben wird wird der Datensatz zu Zahlen transformiert.
}
\examples{


}
\author{
Wolfgang Peter
}
