% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melt2.R
\name{Long}
\alias{Long}
\alias{Long.formula}
\alias{Long.data.frame}
\alias{Long.list}
\alias{Wide}
\alias{Melt2}
\alias{Melt2.formula}
\alias{Melt2.data.frame}
\alias{Melt2.default}
\alias{melt2}
\title{Umformen (reshape)}
\usage{
Long(x, ...)

\method{Long}{formula}(x, .data, ...)

\method{Long}{data.frame}(.data, ..., id.vars = NULL)

\method{Long}{list}(x, .data, by = NULL, key = NULL, value = NULL,
  key.levels = NULL, ...)

Wide(.data, key, value)

Melt2(x, ...)

\method{Melt2}{formula}(x, data, key = "variable", value = "value",
  na.action = na.pass, X = stp25formula::prepare_data2(x, data,
  na.action = na.action), id.vars = X$group.vars, ...)

\method{Melt2}{data.frame}(x, ..., key = "variable", value = "value")

\method{Melt2}{default}(data, ..., key = "variable", value = "value")

melt2(x, ..., by = NULL, key = "variable", value = "value")
}
\arguments{
\item{x}{Objekt kann Formel oder data.frame sein}

\item{...}{weitere Argument}

\item{id.vars}{nur bei Methode data.frame zu verwenden sonst ist hier nichts zu veraendern}

\item{by}{Gruppierung}

\item{key}{Bezeichnung der Bezeichner-Variable default ist "variable"}

\item{value}{Bezeichnung der Werte-Variable default ist "value"}

\item{key.levels}{wenn value gesetzt wird dann 1:nlevels}

\item{data}{Daten}

\item{na.action}{auch nicht zu veraendern}

\item{X}{Formula-Objekt nicht ändern}

\item{...}{weitere Argumente an melt}
}
\value{
data.frame

Dataframe in Langfor
}
\description{
Umformen von einem Breit-Format nach einem Lang-Format. Melt2 und melt2 sind
Erweiterungen der reshape2::melt Funktion. Intern wird  melt und dcast verwendet.

Transformiert von Long nach Wide
Quelle: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378

WArnung: die Funktion ist experimental

Melt2.formula: Melt2(chol0+chol1+chol6+chol12~g, hyper)

Melt2.data.frame ist wie melt zu verwenden aber die Labels werden dabei benuzt.

Melt2.default gibt alles ausser die data.frames und Formeln an reshape2::melt weiter.

melt2 ist die lazy_dots-Methode
}
\examples{
  suppressPackageStartupMessages(library(tidyverse))

df <- data.frame(month=rep(1:3,2),
                 student=rep(c("Amy", "Bob"), each=3),
                 A=c(9, 7, 6, 8, 6, 9),
                 B=c(6, 7, 8, 5, 6, 7))
df \%>\% Wide(student, c(A, B))

df[-4] \%>\% tidyr::spread(student, A)
df[-4] \%>\% Wide(student, A)


df_widw \%>\% Long( Amy_A, Amy_B, Bob_A, Bob_B, by=~month)
Long(list( A=c("Amy_A", "Bob_A"), B=c( "Amy_B", "Bob_B")),
     df_widw,
     by=~month,
     key = "student",
     key.level=c("Amy", "Bob")
     )
 
 t1 <- df \%>\%
 tidyr::gather(variable, value, -(month:student)) \%>\%
   tidyr::unite(temp, student, variable) \%>\%
    tidyr::spread(temp, value)
x<-Melt2(chol0+chol1+chol6+chol12~g , hyper)
aggregate( value~variable, x, mean)

#-- Melt2.data.frame--

x  <- hyper[, c("g","chol0","chol1","chol6","chol12")]
x  <- Melt2(x, id.vars=1)
# aggregate(value~variable+g, x, mean)

# Alternative aber ohne die Labels
x <- hyper  \%>\%
    tidyr::gather("time", "chol", chol0:chol12) \%>\%
    dplyr::select(g, time, chol)


 head(x<-Melt2(chol0+chol1+chol6+chol12~g , hyper))
 # APA2(~.,x)

 head( x<- hyper \%>\%  melt2(chol0,chol1,chol6,chol12, by=~g))
 #APA2(~.,x)
}
