% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XLS.R
\name{XLS}
\alias{XLS}
\title{Nummernindex aus Excel - Spaltenbezeichnung}
\usage{
XLS(...)
}
\arguments{
\item{...}{liste mit den Spaltennamen A:BB}
}
\description{
Extrahiert aus Buchstaben die Spaltennummer
}
\examples{

#strsplit("A:V", "\\:")
XLS(a, B)
XLS(a, B, c:f, g:h,i, r:z)
XLS(A:Z)

}
\author{
Wolfgang Peter
}
